+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+---
+
+## [0.4.0] - 2026-04-11
+
 ### Added
 
 - `MessageType` named type (`int`) for WebSocket frame types
@@ -64,7 +68,8 @@
 - `TextMessage` (1) and `BinaryMessage` (2) untyped int constants for WebSocket frame types
 - `ErrConnectionClosed` and `ErrSendBufferFull` sentinel errors
 
-[Unreleased]: https://github.com/wspulse/core/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/wspulse/core/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/wspulse/core/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/wspulse/core/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/wspulse/core/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/wspulse/core/compare/v0.1.0...v0.2.0